requirements.txt
# Core dependencies
aiohttp>=3.9.1
aiosqlite>=0.19.0
python-dotenv>=1.0.0
watchdog>=3.0.0
redis>=5.0.1
requests>=2.31.0

# Utilities
aiologger>=0.7.0
pydantic>=2.5.0
tenacity>=8.2.3

# API (optional for future web interface)
fastapi>=0.100.0
uvicorn>=0.20.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
